+++ src/components/Feeds/SocialSidebar/Sidebar.jsx
@@ -2,10 +2,10 @@ import React from "react";
 import { SearchContainer } from "../../Explore/ExploreElements";
 import { FilterButton, FilterContainer } from "../FeedsElements";
 import SearchInputBox from "../../Common/SearchInputBox";
+import { filterByTags } from "../../Common/Tags/filterByTags";
 import SocialTags from "../FeedTags/SocialTags";
 import { SidebarContainer } from "./SidebarElements";
 import { RouterNavCreateButtonLink } from "../../Header/Navbar/NavbarElements";
-
 const Sidebar = ({
     user,
     searchTerm,
@@ -20,38 +20,8 @@ const Sidebar = ({
     exploreFiltersComponent,
     sidebarType,
 }) => {
-    const customSplit = (str, invisibleChar) => {
-        const result = [];
-        let temp = "";
-        for (let i = 0; i < str.length; i++) {
-            if (str[i] === invisibleChar) {
-                temp += `${invisibleChar} ${invisibleChar}`;
-                i += 2;
-            } else if (str[i] === " ") {
-                result.push(temp);
-                temp = "";
-            } else {
-                temp += str[i];
-            }
-        }
-        if (temp) result.push(temp);
-        return result;
-    };
-
-    const filterByTag = (tag) => {
-        setSearchTerm("");
-        const invisibleChar = "\u200b"; // Zero width space character
-        const updatedSearchTerm = searchTerm ? customSplit(searchTerm, invisibleChar) : [];
-        if (tag.includes(" ")) {
-            tag = tag.replace(/ /g, `${invisibleChar} ${invisibleChar}`);
-        }
-        const index = updatedSearchTerm.indexOf(tag);
-        if (index !== -1) {
-            updatedSearchTerm.splice(index, 1);
-        } else {
-            updatedSearchTerm.push(tag);
-        }
-        setSearchTerm(updatedSearchTerm.join(" "));
+    const handleFilterByTags = (tag) => {
+        filterByTags(tag, searchTerm, setSearchTerm);
     };
 
     const renderFollowingFilterButtons = () => (
@@ -138,7 +108,7 @@ const Sidebar = ({
 
             {sidebarType === "explore" && exploreFiltersComponent}
 
-            <SocialTags tags={tags} handleClick={filterByTag} />
+            <SocialTags tags={tags} handleClick={handleFilterByTags} />
         </SidebarContainer>
     );
 };
